+++ src/App.tsx
@@ -1,5 +1,5 @@
 import './App.css';
-import Board from './components/Board';
+import Board from './components/Board/Board';
 
 function App() {
   return <Board />;
