+++ src/main.tsx
@@ -1,5 +1,4 @@
 import { createRoot } from 'react-dom/client';
-import './index.css';
 import App from './App.tsx';
 
 createRoot(document.getElementById('root')!).render(<App />);
