+++ src/types.ts
@@ -10,3 +10,10 @@ export type TTask = {
   id: Id;
   title: string; 
 };
+
+export type TPlaceholderProps = {
+  clientHeight?: number;
+  clientWidth?: number;
+  clientX?: number;
+  clientY?: number;
+}
